Cleanup model and persistence classes for notifications
