+++ commons/src/test/resources/application.properties
@@ -1,3 +1,17 @@
 ## Secret Decryption
 #
-%test.secret.key.path=src/test/resources/secret.key
+secret.key.path=src/test/resources/secret.key
+## Hibernate
+#
+# Always use quotes for keywords, column- and table names.
+# e.g. SELECT "FOO"."BAR" FROM "BAZ". This matches what the API server does,
+# and is required for compatibility with its schema.
+quarkus.hibernate-orm.database.globally-quoted-identifiers=true
+# Hibernate should only validate that the existing schema matches our entity classes,
+# but it should never generate a schema by itself.
+quarkus.hibernate-orm.database.generation=validate
+# Use Flyway only in test mode to populate the DB with the schema generated
+# by the API server. In production or dev mode, this will be handled by the
+# API server itself. See https://quarkus.io/guides/flyway
+quarkus.flyway.migrate-at-start=true
+quarkus.flyway.locations=migrations/postgres
