+++ notification-publisher/src/main/java/org/hyades/notification/NotificationRouter.java
@@ -20,15 +20,15 @@ package org.hyades.notification;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.InvalidProtocolBufferException;
-import org.hyades.commonnotification.NotificationScope;
+import org.hyades.notification.model.NotificationScope;
 import org.hyades.exception.PublisherException;
-import org.hyades.model.NotificationPublisher;
-import org.hyades.model.NotificationRule;
+import org.hyades.notification.model.NotificationPublisher;
+import org.hyades.notification.model.NotificationRule;
 import org.hyades.model.Project;
 import org.hyades.model.Team;
 import org.hyades.notification.publisher.Publisher;
 import org.hyades.notification.publisher.SendMailPublisher;
-import org.hyades.persistence.NotificationRuleRepository;
+import org.hyades.notification.persistence.NotificationRuleRepository;
 import org.hyades.persistence.TeamRepository;
 import org.hyades.proto.notification.v1.BomConsumedOrProcessedSubject;
 import org.hyades.proto.notification.v1.NewVulnerabilitySubject;
