+++ notification-publisher/src/test/java/org/hyades/notification/NotificationRouterTest.java
@@ -4,10 +4,10 @@ import com.google.protobuf.Any;
 import io.quarkus.test.TestTransaction;
 import io.quarkus.test.junit.QuarkusMock;
 import io.quarkus.test.junit.QuarkusTest;
-import org.hyades.commonnotification.NotificationGroup;
-import org.hyades.commonnotification.NotificationScope;
-import org.hyades.model.NotificationLevel;
-import org.hyades.model.NotificationRule;
+import org.hyades.notification.model.NotificationGroup;
+import org.hyades.notification.model.NotificationScope;
+import org.hyades.notification.model.NotificationLevel;
+import org.hyades.notification.model.NotificationRule;
 import org.hyades.notification.publisher.ConsolePublisher;
 import org.hyades.proto.notification.v1.BomConsumedOrProcessedSubject;
 import org.hyades.proto.notification.v1.Component;
