+++ notification-publisher/src/test/java/org/hyades/notification/model/NotificationScopeTest.java
@@ -16,31 +16,16 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright (c) Steve Springett. All Rights Reserved.
  */
-package org.hyades.commonnotification;
+package org.hyades.notification.model;
 
-public enum NotificationGroup {
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
-    // System Groups
-    CONFIGURATION,
-    DATASOURCE_MIRRORING,
-    REPOSITORY,
-    INTEGRATION,
-    INDEXING_SERVICE,
-    FILE_SYSTEM,
-    ANALYZER,
+public class NotificationScopeTest {
 
-    // Portfolio Groups
-    NEW_VULNERABILITY,
-    NEW_VULNERABLE_DEPENDENCY,
-    //NEW_OUTDATED_COMPONENT,
-    //FIXED_VULNERABILITY,
-    //FIXED_OUTDATED,
-    //GLOBAL_AUDIT_CHANGE,
-    PROJECT_AUDIT_CHANGE,
-    BOM_CONSUMED,
-    BOM_PROCESSED,
-    VEX_CONSUMED,
-    VEX_PROCESSED,
-    POLICY_VIOLATION,
-    PROJECT_CREATED
+    @Test
+    public void testEnums() {
+        Assertions.assertEquals("SYSTEM", NotificationScope.SYSTEM.name());
+        Assertions.assertEquals("PORTFOLIO", NotificationScope.PORTFOLIO.name());
+    }
 }
